Add ShapeLayer.
