+++ Example/ShapeView/ViewController.swift
@@ -42,7 +42,7 @@ class MessageView: ShapeView {
         }
 
         outerShadow = ShapeShadow(raduis: 20, color: .green)
-        innerShadow = ShapeShadow(raduis: 10, color: .green)
+        innerShadow = ShapeShadow(raduis: 20, color: .green)
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -128,7 +128,6 @@ class ViewController: UIViewController {
         view.backgroundColor = UIColor(white: 1, alpha: 0.5)
         view.blurEffectStyle = .regular
         view.blurAlpha = 0.7
-        
         view.button.setTitle("ShapeView Demo App", for: .normal)
         return view
     }()
@@ -152,8 +151,7 @@ class ViewController: UIViewController {
             $0.addLine(to: CGPoint(x: 0, y: bounds.height))
         }
         shapeView.backgroundColor = UIColor(white: 0.5, alpha: 0.5)
-//        shapeView.shadowColor = .darkGray
-//        shapeView.shadowRadius = 4
+        shapeView.outerShadow = ShapeShadow(raduis: 4, color: .darkGray)
         shapeView.blurEffectStyle = .dark
         shapeView.blurAlpha = 0.8
         return shapeView
@@ -211,12 +209,20 @@ class ViewController: UIViewController {
             $0.left.right.bottom.top.equalToSuperview()
         }
 
+        messageView.snp.makeConstraints {
+            $0.left.equalToSuperview().offset(50)
+            $0.right.equalToSuperview().offset(-50)
+            $0.height.equalTo(80)
+            $0.top.equalToSuperview().offset(70)
+           
+        }
+        
         errorView.snp.makeConstraints {
             $0.centerX.equalTo(messageView)
             $0.size.equalTo(messageView)
-            $0.top.equalToSuperview().offset(70)
+            $0.top.equalTo(messageView.snp.bottom).offset(20)
         }
-        
+
         customView.snp.makeConstraints {
             $0.centerX.equalTo(messageView)
             $0.size.equalTo(messageView)
@@ -227,13 +233,6 @@ class ViewController: UIViewController {
             $0.centerX.equalToSuperview()
             $0.size.equalTo(50)
             $0.top.equalTo(customView.snp.bottom).offset(20)
-        }
-        
-        messageView.snp.makeConstraints {
-            $0.left.equalToSuperview().offset(50)
-            $0.right.equalToSuperview().offset(-50)
-            $0.height.equalTo(80)
-            $0.top.equalTo(starView.snp.bottom).offset(20)
         }
         
     }
